BREAKING CHANGE - fix for A-Frame 1.2.0
A-Frame 1.2.0 uses ThreeJS r125, which removes the Geometry class, which I was using. Rewrote it to use BufferGeometry, somewhat awkwardly. Now works for me in A-Frame 1.2.0 at first glance, but I haven't tested it very thoroughly. so please let me know if you encounter any problems.

To be clear, THIS VERSION OF THE COMPONENT WILL NO LONGER WORK WITH A-FRAME VERSIONS BELOW 1.2.0. However, there are no feature changes, so if you need an older version of A-Frame, you can use the previous version of this component instead.